Remove deprecated gtk_button_set_alignment call
The recommended method is to set the child's `halign` property. However,
because our button has two children that automatically get wrapped in a
GtkAlignment container, we need to set the alignment of that child after
the image and the label have been added.

Also fix an assertion failure when no icon could be found for a file.